[] Add CSS styling to Home page of recipe website
I added styling making the background of the homepage darkgoldenrod.
Also made the links to the recipes no longer have bulletpoints and also
they have a button sort of style to them. This will have to be tweaked
in the near future as im struggling to figure out how to add a noticable
border to them.